Fixed and updated plugin.yml
Implemented large changes to the plugin.yml in an effort to keep in line
with existing standards. Implemented the api-version and load tags to
inform the server runtime about compatibility needs. Also implemented
permissions tags, primarily for popular plugins such as LuckPerms to
correctly complete and provide information on permissions as they are
being modified in the web editor. While adding these permissions tags,
sane default permissions were also defined along with the implementation
of the "servertutorialplus.command.*" tag to enable administrative
commands all in one go. One change to note is that
servertutorialplus.command.quit is now enabled by default for players,
primarily to keep them from trapping themselves inadvertently in a
broken tutorial. IF THIS BEHAVIOR IS NOT DESIRED, NEGATE THE PERMISSION!